requirements.txt
psycopg[binary]
pandas
python-dotenv
fastapi
uvicorn
streamlit
plotly
